Prepare the way for boot-pack.
The scie-jump can now be built with `-p package` which post-processes
the binary to add a 64 bit trailer magic that contains its size. The
scie-jump can use this to cheaply detect if its being run directly as
a stub with no app trailer. This also simplifies the boot-pack by
making the proper binary size known to the binary itself.

In order to do this, switch to nightly to get bindeps support:
https://rust-lang.github.io/rfcs/3028-cargo-binary-dependencies.html

That, in turn led to alot of clippys since format strings are now
enhanced - fix those.