Add support for `{scie}` and `{scie.env.X}`.
This allows natural execution of node entry-point scripts as well as
constructing Python zipapps without mentioning the traling zipapp in
the lift manifest. The zipapp just executes itself via the `{scie}`
placeholder using the embedded Python distribution.

Switch to the placeholders crate from regex to slim down the scie-jump
from 2MB to 1MB and to switch to `{placeholder}` syntax which is
relatively shell friendly unlike `$`-based syntaxes.

Also split off a `context` module from `installer`. There is still work
to do here and the split was messy with clones needing to be sprinkled.

Closes #8