Require name, don't require archive_type.
This paves the way for supplanting the current role for name by an
optional key and turning the name into the file name relative path of
the packed file. This will allow for simpler boot-pack configuration
and support the upcoming need for Directory in addition to the existing
Archive and Blob File variants.