Fix blob locks.
Previously the blob parent directory was locked instead of the lob
itself and this could lead to the blob not being re-unpacked if it
was deleted after 1st unpack.

A previously failing test of this is added to the python example where
the Pants binding command does, in fact, delete the pants.pex blob.