Support files that are lazy loaded at runtime.
Files can now either be embedded in a scie or else loaded at runtime via
a binding cmd that takes the filename to load as an agument and produces
the file content on stdout. Hashing and size checking are still
performed by the installer.

And example is added that assumes `curl`, `jq` and `bash` are present on
the host for doing loading from the internet.

Closes #19