Add support for scie base value placeholders.
Both the lift manifest `scie.lift.base` field and the `SCIE_BASE` env
var now have their values expanded, accepting all placeholders except
for the `{scie.lift}` which depends on the fully evaluated value of the
scie base.

In order to make OS-convention-friendly values feasible, a new
`{scie.user.cache_dir=<fallback>}` placeholder is added that expands to
the OS-appropriate user cache dir when defined and `<fallback>` when
not.