Provide help for the bare `scie-jump`. (#285)
There is now a `-h` / `--help` option as well as a default action of
printing help when no CWD-local lift.json manifest is found.

Also support `-V` / `--version` and fix the scie-jump version reported
in the lift manifest to be the `scie-jump` binary version and not the
jump crate version.